Fix exec bug
I did a small refactor and also fixed an exec bug where we were sometimes getting output all weird.  Turned out we were just getting messages out of order, so I added a queue that we can add our messages to when we get them out of order.
