+++ bctl-go-daemon/bzerolib/stream/stdreader/stdreader.go
@@ -11,6 +11,7 @@ var (
 	EndStreamBytes = []byte{0x62, 0x61, 0x73, 0x74, 0x69, 0x6f, 0x6e, 0x7a, 0x65, 0x72, 0x6f}
 )
 
+// Stdin
 type StdReader struct {
 	StreamType   smsg.StreamType
 	RequestId    string
