+++ bctl-go-daemon/bzerolib/stream/stdwriter/stdwriter.go
@@ -1,6 +1,9 @@
 package stdwriter
 
 import (
+	"encoding/base64"
+	"log"
+
 	smsg "bastionzero.com/bctl/v1/bzerolib/stream/message"
 )
 
@@ -16,6 +19,7 @@ type StdWriter struct {
 	logId          string
 }
 
+// Stdout or Stderr
 func NewStdWriter(streamType smsg.StreamType, ch chan smsg.StreamMessage, requestId string, logId string) *StdWriter {
 	return &StdWriter{
 		StdType:        streamType,
@@ -27,11 +31,13 @@ func NewStdWriter(streamType smsg.StreamType, ch chan smsg.StreamMessage, reques
 }
 
 func (w *StdWriter) Write(p []byte) (int, error) {
+	str := base64.StdEncoding.EncodeToString(p)
+	log.Printf("BYTES AT ENCODING: %v", str)
 	message := smsg.StreamMessage{
 		Type:           string(w.StdType),
 		RequestId:      w.RequestId,
 		SequenceNumber: w.SequenceNumber,
-		Content:        p,
+		Content:        str,
 		LogId:          w.logId,
 	}
 	w.outputChannel <- message
